\name{gpuRNN-package}
\alias{gpuRNN-package}
\alias{gpuRNN}
\docType{package}
\title{
\packageTitle{gpuRNN}
}
\description{
\packageDescription{gpuRNN}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gpuRNN}
\packageIndices{gpuRNN}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gpuRNN}

Maintainer: \packageMaintainer{gpuRNN}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
